Move logger into zeroconf.logger
